# public/swagger.json
{
	"components": {
...
		"schemas": {
...
			"Preference": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"cpf": {
						"type": "string"
					},
					"response": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"Rating": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"score": {
						"type": "number",
						"format": "double"
					},
					"idAccommodation": {
						"type": "number",
						"format": "double"
					},
					"cpf": {
						"type": "string"
					},
					"comment": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Category": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"idPreference": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"keyWord": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Client": {
				"properties": {
					"email": {
						"type": "string"
					},
					"cpf": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Host": {
				"properties": {
					"email": {
						"type": "string"
					},
					"cnpj": {
						"type": "string"
					},
					"fantasyName": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Comfort": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"idPreference": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Complaint": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"idAccomodation": {
						"type": "number",
						"format": "double"
					},
					"cpf": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"AccessGroup": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"History": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "string"
					},
					"dateTimeUpdate": {
						"type": "string",
						"format": "date-time"
					},
					"idBooking": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"Hosting": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"hostType": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"cnpj": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"PermissionType": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Offer": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					},
					"numberOfAdults": {
						"type": "number",
						"format": "double"
					},
					"numberOfKids": {
						"type": "number",
						"format": "double"
					},
					"numberOfRooms": {
						"type": "number",
						"format": "double"
					},
					"numberOfBeds": {
						"type": "number",
						"format": "double"
					},
					"area": {
						"type": "number",
						"format": "double"
					},
					"value": {
						"type": "number",
						"format": "double"
					},
					"idAccomodation": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Reservation": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"cpf": {
						"type": "string"
					},
					"idOffer": {
						"type": "number",
						"format": "double"
					},
					"beginDate": {
						"type": "string",
						"format": "date-time"
					},
					"endDate": {
						"type": "string",
						"format": "date-time"
					}
				},
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "AffectOn API",
		"version": "1.0.0",
		"license": {
			"name": "Gabriela Ribeiro, Joana Mespaque & Louise Queiroz"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/preferencia/id/{id}": {
			"put": {
				"operationId": "Update",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Preferências"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Preference"
							}
						}
					}
				}
			},
			"patch": {
				"operationId": "UpdateAllFields",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Preferências"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Preference"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "FindById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Preference"
								}
							}
						}
					}
				},
				"tags": [
					"Preferências"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Preferências"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/avaliacao/id/{id}": {
			"put": {
				"operationId": "Update",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Avaliações"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Rating"
							}
						}
					}
				}
			},
			"patch": {
				"operationId": "UpdateAllFields",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Avaliações"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Rating"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "FindById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Rating"
								}
							}
						}
					}
				},
				"tags": [
					"Avaliações"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Avaliações"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/categoria/id/{id}": {
			"put": {
				"operationId": "Update",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Categorias"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Category"
							}
						}
					}
				}
			},
			"patch": {
				"operationId": "UpdateAllFields",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Categorias"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Category"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "FindById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Category"
								}
							}
						}
					}
				},
				"tags": [
					"Categorias"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Categorias"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/cliente/email/{email}": {
			"put": {
				"operationId": "Update",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Clientes"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "email",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Client"
							}
						}
					}
				}
			},
			"patch": {
				"operationId": "UpdateAllFields",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Clientes"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "email",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Client"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "FindById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Client"
								}
							}
						}
					}
				},
				"tags": [
					"Clientes"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "email",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Clientes"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "email",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/hospedeiro/email/{email}": {
			"put": {
				"operationId": "Update",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Hospedeiros"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "email",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Host"
							}
						}
					}
				}
			},
			"patch": {
				"operationId": "UpdateAllFields",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Hospedeiros"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "email",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Host"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "FindById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Host"
								}
							}
						}
					}
				},
				"tags": [
					"Hospedeiros"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "email",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Hospedeiros"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "email",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/comodidade/id/{id}": {
			"put": {
				"operationId": "Update",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Comodidades"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Comfort"
							}
						}
					}
				}
			},
			"patch": {
				"operationId": "UpdateAllFields",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Comodidades"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Comfort"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "FindById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Comfort"
								}
							}
						}
					}
				},
				"tags": [
					"Comodidades"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Comodidades"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/denuncia/id/{id}": {
			"put": {
				"operationId": "Update",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Denuncias"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Complaint"
							}
						}
					}
				}
			},
			"patch": {
				"operationId": "UpdateAllFields",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Denuncias"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Complaint"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "FindById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Complaint"
								}
							}
						}
					}
				},
				"tags": [
					"Denuncias"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Denuncias"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/grupo-de-acesso/id/{id}": {
			"patch": {
				"operationId": "UpdateAllFields",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Grupo de Acessos"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AccessGroup"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "FindById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AccessGroup"
								}
							}
						}
					}
				},
				"tags": [
					"Grupo de Acessos"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Grupo de Acessos"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/historico/id/{id}": {
			"put": {
				"operationId": "Update",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Históricos"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/History"
							}
						}
					}
				}
			},
			"patch": {
				"operationId": "UpdateAllFields",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Históricos"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/History"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "FindById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/History"
								}
							}
						}
					}
				},
				"tags": [
					"Históricos"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Históricos"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/local-de-hospedagem/id/{id}": {
			"put": {
				"operationId": "Update",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Locais de Hospedagem"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Hosting"
							}
						}
					}
				}
			},
			"patch": {
				"operationId": "UpdateAllFields",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Locais de Hospedagem"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Hosting"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "FindById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Hosting"
								}
							}
						}
					}
				},
				"tags": [
					"Locais de Hospedagem"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Locais de Hospedagem"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/tipo-de-permissao/id/{id}": {
			"patch": {
				"operationId": "UpdateAllFields",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Tipos de Permissão"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PermissionType"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "FindById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PermissionType"
								}
							}
						}
					}
				},
				"tags": [
					"Tipos de Permissão"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Tipos de Permissão"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/oferta/hospedeiro/{email}": {
			"get": {
				"operationId": "FindOffersByHostEmailRoute",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Offer"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Ofertas"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "email",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/oferta/id/{id}": {
			"put": {
				"operationId": "Update",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Ofertas"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Offer"
							}
						}
					}
				}
			},
			"patch": {
				"operationId": "UpdateAllFields",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Ofertas"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Offer"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "FindById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Offer"
								}
							}
						}
					}
				},
				"tags": [
					"Ofertas"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Ofertas"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/oferta": {
			"post": {
				"operationId": "Create",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Ofertas"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Offer"
							}
						}
					}
				}
			}
		},
...
		"/reserva/id/{id}": {
			"put": {
				"operationId": "Update",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Reservas"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Reservation"
							}
						}
					}
				}
			},
			"patch": {
				"operationId": "UpdateAllFields",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Reservas"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Reservation"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "FindById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Reservation"
								}
							}
						}
					}
				},
				"tags": [
					"Reservas"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Reservas"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}